Refactored Components & Styles
