+++ client/src/components/draw/DrawGrid.jsx
@@ -1,46 +1,17 @@
 import React from 'react';
-import ToggleButton from "../ToggleButton";
-import { MdClose } from "react-icons/md";
-import { ContainerDraw, MainH1, TextLeft, FormAdd, InputText, BtnAdd, SortArea, BtnSort, SortList, ListaMemb } from "./styles.draw";
-import useMemberFunctions from './useMemberFunctions';
+import ToggleButton from "../toggle=button/ToggleButton"
+import { ContainerDraw, MainH1, TextLeft } from "..//../styles/styles.draw";
+import Form from './Form';
+import Grid from './Grid';
 
-function DrawGrid({ toggleTheme, setPairs }) {
-    const { inputValue, setInputValue, members, addMember, deleteMember, sortPairs } = useMemberFunctions();
-
+function DrawGrid({ toggleTheme }) {
     return (
         <ContainerDraw>
             <ToggleButton toggleTheme={toggleTheme} />
             <MainH1>Calendário de Limpeza</MainH1>
             <TextLeft>Insira o nome</TextLeft>
-            <FormAdd onSubmit={(e) => {
-                e.preventDefault();
-                addMember(inputValue);
-            }}>
-                <InputText
-                    placeholder='Ex. João'
-                    value={inputValue}
-                    onChange={(e) => setInputValue(e.target.value)}
-                    onKeyPress={(e) => {
-                        if (e.key === 'Enter') {
-                            addMember(inputValue);
-                        }
-                    }}
-                />
-                <BtnAdd onClick={() => addMember(inputValue)}>Adicionar</BtnAdd>
-            </FormAdd>
-
-            <SortArea>
-                <TextLeft>Lista de participantes</TextLeft>
-                <BtnSort onClick={() => sortPairs(setPairs)}>Sortear</BtnSort>
-            </SortArea>
-            <SortList>
-                {members.map((member, index) => (
-                    <ListaMemb key={index}>
-                        {member.name}
-                        <MdClose onClick={() => deleteMember(index)} />
-                    </ListaMemb>
-                ))}
-            </SortList>
+            <Form />
+            <Grid />
         </ContainerDraw>
     );
 }
