+++ client/src/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import App from './App';
-import { GlobalStyle } from "./components/GlobalStyle";
+import { GlobalStyle } from "./styles/GlobalStyle";
 
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
