+++ client/src/styles/styles.draw.js
@@ -6,62 +6,7 @@ export const ContainerDraw = styled.div`
     width: 80%;
     margin-top: 8%;
 `;
-export const SortArea = styled.div`
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-    padding: 10% 0% 10% 0%;
-    width: 90%;
-`;
-export const SortList = styled.ul`
-    overflow: auto; 
-    margin-bottom: 8%;
-    height: 55vh;
-    color: ${(props) => props.theme.withe};
-`;
 
-
-
-export const InputText = styled.input`
-    display: flex;
-    justify-content: center;
-    margin: 4% 0% 8% 0%;
-    border: solid 1px;
-    border-radius: 5px;
-    height: 4vh;
-    width: 85%;
-`;
-export const FormAdd = styled.div`
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    color: ${(props) => props.theme.withe};
-`;
-export const BtnAdd = styled.button`
-    background-color: #B73625;
-    font-family: 'Montserrat';
-    cursor: pointer;
-    color: white;
-    border: none;
-    padding: 4px;
-    border-radius: 5px;
-    width: 100px;
-    &:hover {
-        box-shadow: 4px 4px 0px 0px #772419;
-  };
-`;
-export const BtnSort = styled.button`
-    font-family: 'Montserrat';
-    background-color: #B73625;
-    cursor: pointer;
-    color: white;
-    border: none;
-    padding: 4px;
-    border-radius: 5px;
-    &:hover {
-        box-shadow: 4px 4px 0px 0px #772419;
-  };
-`;
 export const MainH1 = styled.h1`
     font-family: 'Montserrat';
     display: flex;
@@ -77,6 +22,68 @@ export const TextLeft = styled.p`
     font-size: 15px;
     color: ${(props) => props.theme.withe};
 `;
+
+export const FormAdd = styled.div`
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    color: ${(props) => props.theme.withe};
+`;
+
+export const InputText = styled.input`
+    display: flex;
+    justify-content: center;
+    margin: 4% 0% 8% 0%;
+    border: solid 1px;
+    border-radius: 5px;
+    height: 4vh;
+    width: 85%;
+`;
+
+export const BtnAdd = styled.button`
+    background-color: #B73625;
+    font-family: 'Montserrat';
+    cursor: pointer;
+    color: white;
+    border: none;
+    padding: 4px;
+    border-radius: 5px;
+    width: 100px;
+    &:hover {
+        box-shadow: 4px 4px 0px 0px #772419;
+  };
+`;
+
+export const Grid = styled.div`
+`;
+
+export const SortArea = styled.div`
+    display: flex;
+    flex-direction: row;
+    justify-content: space-between;
+    padding: 10% 0% 10% 0%;
+    width: 90%;
+`;
+export const BtnSort = styled.button`
+    font-family: 'Montserrat';
+    background-color: #B73625;
+    cursor: pointer;
+    color: white;
+    border: none;
+    padding: 4px;
+    border-radius: 5px;
+    &:hover {
+        box-shadow: 4px 4px 0px 0px #772419;
+  };
+`;
+
+export const SortList = styled.ul`
+    overflow: auto; 
+    margin-bottom: 8%;
+    height: 55vh;
+    color: ${(props) => props.theme.withe};
+`;
+
 export const ListaMemb = styled.li`
     display:flex;
     flex-direction: row;
